refactor(ValidationTest): 생성자 검증으로 테스트 리팩토링
